+++ tests/DependencyInjection/Compiler/Fixtures/OtherExampleEvent.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\Ddd\Tests\DependencyInjection\Compiler\Fixtures;
+
+class OtherExampleEvent extends ExampleEvent
+{
+}